[neutron-agents] Remove liveness probes from agents
Current liveness probes ask the neutron server if there was an
heartbeat received from itself. This has a dependency on the neutron
server. In consequence, if the neutron-server is overloaded the pods
will be rolled because the liveness probe failed. This in turn leads to
some pods executing heavy startup query (+ the liveness probes that are
executed more frequently due to failures) adding to the neutron-servers
load that is already high. For the time being, these must be removed to
reduce load on the neutron-servers. Broken agents are still covered by

https://github.com/sapcc/helm-charts/blob/HEAD/openstack/neutron/alerts/neutron.alerts#L4-L18

In the long run we should thing about exposing an http endpoint on
`/healtcheck` that returns 200 if we can still reach rabbitmq and other
things.