[neutron] Options to indicate and configure decomssioning or no schedule
We introduced to disable scheduling on a particular agent in the asr1k
code. This feature can be used either if an agent is supposed to be
decomissioned or if we do not want to increase the load on an agent.

However that is an important difference for the capacity calculation,
hence we employ different annotations on the pod, depending of the
purpose. If an agent is to be torn down and never used again,
`decomissioning` must be used. It indicates that the agent should not ne
included in the capacity calculation.

If we just intend to prevent any additional load on the agent,
`scheduling_disabled` must be used. The agent shall than still be
considered as capacity available.